feat: improve performance on both players
