+++ src/app/components/fullscreen/player.tsx
@@ -7,29 +7,20 @@ import {
   SkipBack,
   SkipForward,
 } from 'lucide-react'
-import { useCallback, useState } from 'react'
 import { Button } from '@/app/components/ui/button'
-import { Slider } from '@/app/components/ui/slider'
 import {
   usePlayerActions,
   usePlayerIsPlaying,
-  usePlayerProgress,
   usePlayerState,
 } from '@/store/player.store'
-import { convertSecondsToTime } from '@/utils/convertSecondsToTime'
 import { LikeButton } from './like-button'
+import { FullscreenProgress } from './progress'
 import { VolumeContainer } from './volume-container'
 
-let isSeeking = false
-
 export function FullscreenPlayer() {
-  const [localProgress, setLocalProgress] = useState(0)
-  const progress = usePlayerProgress()
   const isPlaying = usePlayerIsPlaying()
-  const { audioPlayerRef, currentDuration, isShuffleActive, isLoopActive } =
-    usePlayerState()
+  const { isShuffleActive, isLoopActive } = usePlayerState()
   const {
-    setProgress,
     isPlayingOneSong,
     toggleShuffle,
     playNextSong,
@@ -40,46 +31,10 @@ export function FullscreenPlayer() {
     toggleLoop,
   } = usePlayerActions()
 
-  const handleSeeking = useCallback((amount: number) => {
-    isSeeking = true
-    setLocalProgress(amount)
-  }, [])
-
-  const handleSeeked = useCallback(
-    (amount: number) => {
-      isSeeking = false
-
-      if (audioPlayerRef) {
-        audioPlayerRef.currentTime = amount
-        setLocalProgress(amount)
-        setProgress(amount)
-      }
-    },
-    [audioPlayerRef, setProgress],
-  )
-
   return (
     <div className="w-full">
-      <div className="flex gap-4 items-center">
-        <div className="min-w-[50px] text-right">
-          {convertSecondsToTime(progress)}
-        </div>
+      <FullscreenProgress />
 
-        <Slider
-          variant="secondary"
-          defaultValue={[localProgress]}
-          value={isSeeking ? [localProgress] : [progress]}
-          max={currentDuration}
-          step={1}
-          className="w-full h-4"
-          onValueChange={([value]) => handleSeeking(value)}
-          onValueCommit={([value]) => handleSeeked(value)}
-        />
-
-        <div className="min-w-[50px] text-left">
-          {convertSecondsToTime(currentDuration ?? 0)}
-        </div>
-      </div>
       <div className="flex items-center justify-between gap-4 mt-5">
         <div className="w-[200px] flex justify-start">
           <LikeButton className={buttonStyle.secondary} />
